Trigger email notification settings dialog from anywhere
